# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TICKET API",
        "description": "Documentation de mon API Laravel",
        "contact": {
            "name": "Support API",
            "email": "shadracmeshacdimon@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Récupérer toutes les commandes",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Liste des commandes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Créer une nouvelle commande",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_number",
                                    "order_event_id",
                                    "order_price",
                                    "order_type",
                                    "order_payment"
                                ],
                                "properties": {
                                    "order_number": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "order_event_id": {
                                        "type": "integer"
                                    },
                                    "order_price": {
                                        "type": "integer"
                                    },
                                    "order_type": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "order_payment": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "order_info": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Récupérer une commande par ID",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/orders/number/{orderNumber}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Récupérer une commande par numéro de commande",
                "operationId": "491ac82e8a5d57333691fde4c9e6cf48",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "path",
                        "description": "Numéro de commande",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commande non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/user/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Récupérer les commandes de l'utilisateur",
                "description": "Récupère une liste paginée des commandes de l'utilisateur authentifié",
                "operationId": "30079d06b088ad990e27d9f5f092be4c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des commandes de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Modèle représentant une commande",
                "properties": {
                    "order_id": {
                        "description": "ID unique de la commande",
                        "type": "integer"
                    },
                    "order_number": {
                        "description": "Numéro unique de la commande",
                        "type": "string"
                    },
                    "order_event_id": {
                        "description": "ID de l'événement associé à la commande",
                        "type": "integer"
                    },
                    "order_price": {
                        "description": "Prix de la commande",
                        "type": "integer"
                    },
                    "order_type": {
                        "description": "Type de commande",
                        "type": "string"
                    },
                    "order_payment": {
                        "description": "Méthode de paiement utilisée pour la commande",
                        "type": "string"
                    },
                    "order_info": {
                        "description": "Informations supplémentaires sur la commande",
                        "type": "string",
                        "nullable": true
                    },
                    "api_key": {
                        "description": "Clé API utilisée pour créer cette commande",
                        "type": "string"
                    },
                    "order_created_on": {
                        "description": "Date et heure de création de la commande",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date et heure de la dernière mise à jour de la commande",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Use the token provided after logging in.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}